projet_semaine_rafal/requirements.txt
altair==5.3.0
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docutils==0.20.1
extra-streamlit-components==0.1.71
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
packaging==24.1
pages==0.3
pandas==2.0.3
piglet-templates==1.3.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==5.27.2
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
six==1.16.0
smmap==5.0.1
st-theme==1.2.3
streamlit==1.36.0
streamlit-authenticator==0.3.2
streamlit-calendar==1.2.0
streamlit-navigation-bar==3.3.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zipp==3.19.2
